+++ examples/nest/with-emailpassword/backend/src/app.module.ts
@@ -1,0 +1,19 @@
+import { Module } from '@nestjs/common';
+import { AppController } from './app.controller';
+import { AppService } from './app.service';
+import { AuthModule } from './auth/auth.module';
+import * as SuperTokensConfig from './config';
+
+@Module({
+  imports: [
+    AuthModule.forRoot({
+      // try.supertokens.com is for demo purposes. Replace this with the address of your core instance (sign up on supertokens.com), or self host a core.
+      connectionURI: SuperTokensConfig.connectionUri,
+      // apiKey: "IF YOU HAVE AN API KEY FOR THE CORE, ADD IT HERE",
+      appInfo: SuperTokensConfig.appInfo,
+    }),
+  ],
+  controllers: [AppController],
+  providers: [AppService],
+})
+export class AppModule {}