+++ examples/nest/with-emailpassword/backend/src/auth/auth.filter.spec.ts
@@ -1,0 +1,7 @@
+import { AuthFilter } from './auth.filter';
+
+describe('AuthFilter', () => {
+  it('should be defined', () => {
+    expect(new AuthFilter()).toBeDefined();
+  });
+});