+++ lib/ts/nextjs.ts
@@ -39,7 +39,6 @@ export default class NextJS {
         response: any
     ): Promise<T> {
         return new Promise<T>(async (resolve: any, reject: any) => {
-            request.__supertokensFromNextJS = true;
             try {
                 let callbackCalled = false;
                 const result = await middleware((err) => {